Protect JID cache access with mutex.
The overhead appears to be neglectible.

Using TSS had a similar overhead, with the disadvantages that it uses
more storage (one cache per thread vs one cache), has less cache reuse,
and that Windows/MSVC doesn't support automatic TSS cleanup, so it
requires manual cleanup work (which is non-trivial).

Note that the mutex approach may yield more overhead in multi-threaded
applications (in case of contention). Currently, the mutex also locks
during the whole nameprep sequence, which is not strictly necessary.