Introduce system settings policies.
Release-Notes: It is now possible for sysadmins to deploy files with policies for configuration options, such as making it impossible for users to save passwords or to force sound notifications off, or to set defaults.

Also allow changing an option so that Swift disconnects on idle timeout, instead of going away.