Don't crash if server doesn't send cert in TLS handshake
This patch reflects a change to Stroke, which is described as follows:

If a TLS connection results in the server choosing an anonymous cipher
suite, then no server certificate will be returned by the server.

This ought not to happen, since XMPP clients are expected only to
propose non-anonymous cipher suites, but it could be that a client is
coded to propose anonymous suites, or that a bug in the server means
that it fails to return a server certificate.

This change updates the ServerIdentityVerifier to make it resilient
against these situations, treating this situation as equivalent to
"certificate presented by server does not verify".

Test-information:

Tested in Stroke. Untested in Swiften

Change-Id: Iec815b09b6be675edad1d479d1a0a9d6b0b91bf3