Fix attempting to start impromptu chat when impromptu not supported.
Test-Information:

Connect to a server where impromptu is NOT supported. Confirm that adding multiple users by dragging or typing into the JID box is not allowed.
Connect to a server where impromptu is supported. Confirm that it is still possible to drag multiple users and add multiple users by typing into the JID box.

Change-Id: Ibb4c1f7d6a5820f7ccf3629c1e0d30a441d9d159