Hide QtSuggestingJIDInput suggestions when the list is empty.
Test-Information:

Try search terms that return results and verify that the suggestions are shown. Try search terms that return no results and verify that the suggestions list is hidden. Verify that focus-out still hides the suggestion list correctly and that selecting items from the suggestion list still adds the JID to the contact list like before.

Change-Id: I98542df578da9bcfe8d545c77da9d82d6e69a2a6