Reset roster filter when hitting enter to start chat.
Test-Information:

Enter search term and use keyboard arrows to move to select a contact and pressing enter will start a chat and clear the filter. Confirm that pressing escape still clears the filter without starting a chat and that the changes do not interfere with starting a chat normally by double clicking on a contact.

Change-Id: I90f5d431da56896eeb10f16c0ba23bdc143c4857