Fix recent chat items not being shown for private MUC messages.
Test-Information:

Send private message and verify that the private message item in the recents list is erased when the user leaves the MUC and the chat window is closed. Check that other recent items are not removed. Check that private message recent items are not saved and loaded when the application is restarted.

Change-Id: I62b9d324143d2e77ed98592cf37fb681165285c2