Show status message instead of empty VCard dialog when receiving VCard or when an empty VCard is received.
Test-Information:

Request VCard from a user in a MUC that is configured to NOT allow VCards from anonymous users and observe that "empty vcard" status message is displayed instead of an empty dialog. Check that VCard is displayed properly when a VCard is available and that the throbber is still functioning as before.

Change-Id: I098b3ad9495b06b4efbca1547021311f5205cbf2