Fix displaying blocked user warning with bound JID chats. Support multiple alert messages.
Test-Information:

Block a user while the JID is unbound and verify that the yellow warning bar is displayed. Send bidirectional messages to bind the JID then repeat the block request and confirm that the yellow warning bar is displayed and that the user is indeed blocked. Open several alert messages, check that each one can be individually closed via cancelAlert.

Change-Id: I120e393c028d15fd5f92154b44a8817cbc41edc9