Allow injection of flags into unit test checker
Some projects (e.g. Eve-Xin) piggy-back on Swift's build system, but are
unable to do the same with unit tests until this patch (because the checker
will need to link against the unknown library to test it). This patch
allows them to set INJECTED_UNITTEST_FLAGS, which will get included in the
checker's compilation if set.

Test-Information:

Added unit test to Eve-Xin, which builds and runs.

Change-Id: I58ffb2043b968144ad10096b83ac9d88f54f72f3