Hide unchecked preferred stars in profile editor where there is only one field.
Hide the preferred stars in profile editor if it is not checked and there is
only one field of a field type in the vCard. Example: If you have only one eMail
in your vCard it does not make much sense to mark it preferred. If it is already
marked as preferred we will still display it as such.

Test-Information:
Manually tested it via editing a vCard of a test account. On addition or deletion
of fields other fields are checked and stars hidden/shown on demand.

Change-Id: I4704d52518e662f7e3a168ed2b42602383b2153f