Fix own avatar in contact list header not updated on VCard edit.
Commit 14fd8e changed the JID comparison to JID::WithResource. However, we want
JID::WithoutResource because a) handleAvatarChanged is called with a bare resource
when our own avatar changes and b) we also want the avatar in the contact list
header be updated if another of our resources changes it.

Test-Information:
Changed own avatar to a different avatar or cleared it. Before the avatar in the
contact list header was not updated; now it is.

Change-Id: Ie84b6cb10b528f75c1c76c7ceae558f00380d99f