Fix QLayout related warning.
QVBoxLayout was initialized with a widget which automatically tries to make the
new layout the main layout of the widget. However, this is not intended. This bug
was introduced in cc379dc.

Test-Information:
Before a warning "QLayout: Attempting to add QLayout..." was printed to stdout
when joining a MUC; now this warning is gone.

Change-Id: Ie680b188e108a3d825fb98947f2829d3c12955f7