Only handle impromptu MUC invites once.
Added missing return-statement after handling the conversion from 1-to-1 to MUC.
This lead to an additional tab being opened.
In addition the indentation of the relevant block was wrong and is fixed now.

Test-Information:

Before there was a reproducible scenario: test@A starts to chat with test2@A.
toast@A is on test@A's roster, but not on test2@A's roster. test@A drops toast@A
on the running chat between test@A and test2@A. test@A and test2@A end up with
the chat converted into a MUC. toast@A automatically joins the MUC. test2@A
additionally has a new tab.
Now test2@A does not have the additional tab anymore. Only the MUC tab which
previously was a 1-to-1 tab.

Change-Id: I32045812952c840f0f9c6da3a9aecf5c63797b64