Limit NAT-PMP detection timeout to 10 seconds.
libnatpmp has a high (over 2 minutes) timeout for discovery of NAT-PMP
support. This leads to a bad user experience when there is no NAT-PMP
support in the network.

Test-Information:

Tested this in an university network where neither UPnP nor NAT-PMP is
available. File-transfers waited more than 2 minutes for an answer of
a NAT-PMP device before starting the file-transfer.

Change-Id: I3ceaac7f89c982ffa683ee367d5ea3c3d4a2c5ca