Fix shortcut handling for chat window navigation
This commit fixes CTRL-Tab, CTRL-Backtab and CTRL-W shortcuts on Mac,
Windows and Linux.

These shortcuts work in all mode except --no-tabs mode. This will be fixed
in a future commit. There is still different behavior across platforms:
on Windows and Linux they are handled on key press and repeat, and on
Mac they are handled on key release and do not repeat.

Test-Information:

Tested on OS X 10.9.5, Windows 7 and Elementary OS.

Change-Id: I2179f234cfd76b86cf261f2cf7500fc1cf0439a4