Fix focus handling bug with regard to tab changing
QtChatWindow: Require now parameter is set in QtChatWindow::qAppFocusChanged
which is required so only the visible tab that is changed to gets focus.

QtDynamicGridLayout: A change of tab in a QTabWidget set focus on the new
tab even if the QTabWidget has no keyboard focus.

Test-Information:

Tested on Mac OS X 10.9.5 and Qt 5.4.0.

Change-Id: I3d9d230c7753f1898b251b49c06fe7318ad953fb