Fix ASAN reported use-after-free error
The class connects to SettingsProvider::onSettingChanged in its c-tor
but never disconnects from the signal. When an instance is deleted and
the signal is called afterwards it will call the method which will try
to access class members which have already been deleted.

Test-Information:

Tested on OS X 10.9.5. ASAN reported it sometimes on exit. Have not
been able to reproduc it anymore with this fix.

Change-Id: I7f1d815dca87f84a4ae93a5455307e261a1ee329