Show a warning notice when trying to enter a blocked room
When the user tries to enter a blocked room, we now show a warning
notice and describing how the room can be unblocked. Swift will not
send the joining presence when trying to enter a blocked room.

Test-Information:

Tested on Mac OS X 10.9.5 against a popular open source server and its
MUC and Blocking Command implementation.

Change-Id: I875db056f21f97845c5a9a43167b0f2a16bdaa36