Use windeployqt.exe if available and Win distribution documentation fix
If windeployqt.exe is available for Qt5, it will be used to detect
which Qt DLLs to put in the Windows distribution and its dependencies
correctly.

Added a note to our Windows building documentation about the VS
redistributable.

Test-Information:

Tested building a MSI package on Windows 8.1 Pro with Qt 5.3.2 32-bit
msvc2013_opengl which successfully installs and runs.

Change-Id: I786da40d6467f1de8e64bfae275f8363ac1d5ba8