Fix data race in DummyEventLoop and BoostConnection(Server)Test reported by TSAN
The data race is on the events_ member in DummyEventLoop.

A BoostIOServerThread can post events to the DummyEventLoop and thereby
access its events_ data member while the test's main code processes
events of the loop.

To prevent access to the DummyEventLoop by the BoostIOServiceThread
after the DummyEventLoop is deleted, the BoostIOServiceThread is
deleted before the DummyEventLoop.

Process remaining events in BoostConnectionTest::tearDown like we do in
BoostConnectionServerTest::tearDown.

Test-Information:

Run multiple times on NetworkTest as TSAN enabled build on Linux.
Without the patch it reports a data race for the events_ deque member.

Change-Id: I3c85535338fc0ce0263dbfc3534aceb1dd09c137