Do not try highlight words if no text and background colours are specified
The tests have been adjusted so a colour is set in their highlight
actions.

Test-Information:

Tested with a "own nick" highlight rule with a "No highlight" action
and one with highlights colours set. The cases now work as expected.

Added a unit test that checks that no highlighting happens when no
colour is specified.

Change-Id: Ied48d8c2e033531c1b0532348b9380e1bd884b44