Fix crash when a tab drag is aborted in trellis mode
To prevent an infinite recurison during when a tab is returned (added
again) to the original QTabWidget.

The recursion was caused by a show event that is emmitted during insert
and our QtChatTabs::handleWidgetShown function trying to add the tab
again because it is not in the QTabWidget yet; it would be if the show
event would be emitted after the insert is done and not somewhere
during the insert.

Test-Information:

Dragging a tab in trellis mode and *not* dropping it on a tab bar.
Previously this would cause an infinite recursion. Now it just returns
the tab to its previous place.

Tested on OS X 10.9.5 with Qt 5.4.0.

Change-Id: Icd54b95b6f0542b6a9d3921fa1b0ab4df409950a