Change IQ handling to allow non-final IQ handlers/responders
Previously all IQ handlers were final. The IQ handles were processed
in last-added-first order and if the IQ payload matched the handler
payload, the IQ would only be passed to that handler.

This conflicts with our IBBReceiveSession. When running multiple
concurrent file-transfers using multiple IBBReceiveSessions there are
multiple IQ handlers for IBB content; one for every transfer.

This commit allows a Responder to be set as non-final. In this case
unhandled IQs will not be responded with an error but returned to
the IQRouter so it can pass it to the next possible IQ handler.

Test-Information:

Tested with ConcurrentFileTransferTest with runs multiple IBB transfers
in parallel.

Change-Id: I8237e234cbe5c110deaa8c3d6ba303b65fd53d00