Add resource management for S5B server and port forwardings
This patchs adds management of the SOCKS5 bytestream server and port
forwarding setup to Swiften. The first file-transfer, regardless of
incoming or outgoing transfer, will start and initialize the S5B server
and if configured try to set up port forwarding. The last file-transfer
to finish will will close the server and remove the port forwarding.

Test-Information:

Tested with upcoming ConcurrentFileTransferTest.cpp and running multiple
S5B file-transfers in parallel.

Change-Id: Idd09d3d0ef9498fc9435b0aee81f2b1061783342