Only close unneedded S5B connections on the S5B server if a S5B server is running
Add a check on whether we have a running SOCKS5 bytestream server
before trying to access it to close unneeded S5B connections to it
after we decided on a candidate to use.

Test-Information:

Without this FileTransferTest crashes in the configuration that does
not allow direct or assisted S5B connections. This patch fixes this
issue.

Change-Id: Ifcf59f82755b9620e7f5fa5841e45815a41ac442