Fix for HTTPConnectProxiedConnection to support responses in pieces
This fix lets HTTPConnectProxiedConnection buffer response data in
pieces until the end of the HTTP header is reached. Only then it will
try to parse the HTTP header.

This is *not* the HTTP chunked transfer encoding.

Test-Information:

Adjusted one test to respond in pieces and added a new test case that
verifies that response data is buffered.

Change-Id: Icfb987bdf2fc5771401a8a9c6979fa9ad1eebdca