Fix popup visibility issues in QtSuggestingJIDInput
When hiding the popup after selection of an item via mouse, the focus
was given to a widget other then the QtSuggestingJIDInput or its parent
dialog. This patch explicitly sets the focus back to the
QtSuggestingJIDInput.

Hide the potentially open popup when the QtSuggestingJIDInput is hidden.
This can happen when the user closes a dialog with a QtSuggestingJIDInput
inside of it.

Test-Information:

Verified described behavior on OS X 10.9.5 and Windows 8.

Change-Id: Ic6629e6e626be18a70de159df62cda79dd82ee09