Add emoticons input popup
This is simple popup menu added to chat window where it is possible to choose an emoticon. To activate click on the smiling face next to input field.
When emoticon is selected its text equivalent is appended to input.

License:
This patch is BSD-licensed, see Documentation/Licenses/BSD-simplified.txt for details.

Test-Information:
Tested on Ubuntu 14.10 with KDE and Qt 4.8.6

Change-Id: I6c5907959970398c9c38591f64ceec20efcbf409