Introduce FeatureOracle class for contact feature support detection
This modifies the feature detection in the ChatController to try to use
the common features of all available resources feature detection if no
full JID has been bound to the chat yet.

Test-Information:

Tested with two Swift instances. Tested
a) the initial chat start case and,
b) the offline/online.

In case a) Swift used to initally show a yellow warning about no support
for message receipts. This warning is gone now.

In case b), after a user gone offline and online again in a running chat,
Swift used to show a warning about missing support for message receipts.
This warning is gone now.

Change-Id: I7a769fde8d14847b180503aeaa58280c572d81b3