Disable signals when adding a tab to the trellis layout
Due ot the way Qt implements widgets and the events, the windowOpening
signal of a QtTabbable is emitted before the QtTabbable is a member of
a QTabWidget when adding a QtTabbable to a QTabWidget.
Disabling the signals when adding the widget will prevent these signals
to be called and will prevent and endlress recursion.

This has originally been reported by Pavol Babincak.

Test-Information:

Without this patch the reporter had Swift crash after a day or two.
With this patch the reporter does not experience the endless recursion
(crash) anymore.

Change-Id: I8b3d791d79e9ab8f42dd1a1cba30d118f620f15d