Skip name lookup if manual hostname is a valid IP address
Removed dead debugging code and unnecessary includes as well.

Test-Information:

Tested with Swiften/QA/ClientTest, adjusted to have a manual hostname
set in the ClientOptions.
Tested on OS X 10.9.5, with NDK API level 14 on an emulated Android 4.0
ARM instance.

Without this patch the connector timed out during name lookup trying
to resolve the IP set in the ClientOptions::manualHostname. With this
patch it skips lookup and connects successfully.
Tested and verified unchanged behavior with this patch and without the
manual address override set.

Change-Id: I737327b2e66c9da78a1963e754bcf201b7d40626