Fix trellis related bugs
Bugs like:
* Tab title not matching shown chat window
* Duplicated tab titles after tab movement

Swift is also subject to QTBUG-36455, which is fixed for Qt >= 5.3.0.

This commit removes the use of application wide focus handlers in
QtChatWindow class, the QtChatWindow::qAppFocusChanged method.
The reason for this is due to the way QTabBar::moveTab is implemented
in Qt which we use for the Trellis feature.
Internally QTabBar::moveTab first adjusts its tab bar labels, then
removes the tab from its old location in the internal QStackedLayout
and then inserts it in the new location. After the remove Qt gives
focus to another widget via a focus change that does not go through
the event loop of the application and is not interceptable with
eventFilters.
Previously we would set the focus and call other signals in the
application wide focus change handler which then used the currently
inconsistent QTabBar/QTabWidget.

Test-Information:

Tested tab switching and movement on  Windows 8 (Qt 5.3.2),
OS X 10.9.5 (Qt 5.4.3) and Ubuntu 14.04.2 LTS (Qt 5.4.3).

Change-Id: Ief423c4add58a90279109f72fac95fc58cb71111