Fix SCons qt4 tool pkg-config detection
Correctly use pkg-config to test for Qt, if no qt path is specified
in config.py. 'Import(…)' at the top caused an error which was hidden
by try/catch at upper layer.

Test-Information:

Tested on Kubuntu 14.04, with and w/o qt path set and verified that
no exception is thrown if Qt is registered at pkg-config.

Change-Id: I82800c23e9f75bd5f6b136fe384a70165a92bb45