Show old vCard in UI if vCard update is rejected by server
Previously QtVCardWidget would work on a shared copy of a VCard and
modify the shared version. Now when setting the vCard on the widget
it creates its own copy and works on that.

The ProfileController holds a reference to the VCard it sets on the UI.
If the server rejects an update it can reset the UI to this VCard.

Test-Information:

Tested against a XMPP server that rejects all vCard updates
(mod_readonly).

Change-Id: Iaa8eed6406f2dc6cc7ac5434194a7dae4f879b33