Wait for assisted candidate detection before emitting candidates
Due to missing signal connection and a copy'n'paste bug, Swift did not
wait for port forwarding/public IP detection before emitting the list
of local candidates. This is fixed now.

The signal is automatically disconnected when the file transfer
is finished and s5bServerPortForwardingUser_ is freed.

Test-Information:

Send a file between two Swift instances and verified log output.

Change-Id: I6530a7ac1cbf6941061bd99aa3f3b0624ebc984c