Create S5B proxy candidates for each DNS lookup result
If the S5B proxy lookup result has no direct IP address listed the
name is resolved using DNS. This change will create a proxy result
per resolved result address and not only the first result address.

Test-Information:

Verified the candidate list contains both entires for a proxy name
that resolves to an A and an AAAA record.

Change-Id: Iec21ff90af981030ff49fb53803d88a59694767c