React to file transfer cancellation before the transfer started
Connect FileTransfer::onStateChanged signal earler to slot in
FileTransferController, so the slot will be called for state changes
that occur before a transfer started.

Test-Information:

Tested with two Swift instances and cancelling the sender side and
receiver side. Without this patch the receiver side UI does not change
if the receiver side cancels the transfer.

Change-Id: I1d4d3e1ac78689a16120bfa9ccdec30ab1191ee3