Show collecting of file transfer candidates in UI
Show the user the collection of possible file transfer candidates
after pressing the 'Start' button for a file transfer. Previously the
buttons remained in the UI giving no feedback to the user at all. If
no UPnP/NAT-PMP device is present, it this stage can take a couple
seconds to timeout and move on to the next stage.

Furthermore this commit adds documentation for the different states
in the ChatWindow::FileTransferState enum.

Test-Information:

Tested this in a network environment with no UPnP/NAT-PMP device
between two Swift instances.

Change-Id: I76ec6e641a2acd683938fe2d8f542d023a244145