Remove when sending an empty response
In RFC 6120, there are only 2 places where "=" is allowed - in the initial
client auth and in the final server success response.

While testing challenge response exchanges in Kerberos, it was seen that Swift
was adding an '=' in other empty responses. This patch fixes it by sending an
empty response instead of an '='.

Test-information:
Tested on Windows using a WIP GSSAPI authentication exchange with M-Link.

Change-Id: I1f82bddbd1380361cbe43e45a2804156249582ae