Stay scrolled at bottom of chat view during a file transfer
Swift did not remember the correct scrolled-at-bottom position for
the current chat view after it added/updated file-transfer UI to the
chat view.

Test-Information:

Tested with two Swift instances (one with the fix, one without) and
send files in both directions. The instance with this fix, scrolls
down along with new messages as they are recevied if the chat view
is scrolled to the bottom.

Change-Id: Ia9afa68254c3add5f7677b1f4b7ed3c7e1edcf6f