Use GSSAPI when SSO is used
This patch uses the GSSAPI authenticator on Windows if the server advertises it
and the client requests it. (The user is not able to request it in the UI yet)

Also sends the manual port to the GSSAPI authenticator to construct the SPN if
a non-default port is used.

Test-information:
Tested on Windows using WIP code. Tested both on TLS & without.
Unit tests pass.

Change-Id: I9a9ad9604fe084d5fb2003b7a91174a9512e2eec