Fix memory leak in QtUserSearchWindow
The delegate was not correctly freed and was leaking across
login/logout sessions.

Setting the parent of the delegate during initialization, so it
will be freed by Qt when the QtUserSearchWindow is freed.
Delegated the deallocation to Qt, because it has to be freed
after the widget that uses the delegate is freed.

Test-Information:

The leak was reported by Valgrind. Verified with Valgrind that
this commit fixes the leak and is not reported anymore.

Change-Id: Ib48d2628577433bce82103b3f14cbd11d733b7ac