Do not reallocate FileTransferManager on connect of Client
Now Client allocates a FileTransferManager in its ctor and frees
it in its dtor, like all the other resources of the Client class.

This fixes an ASAN crash in the use case where you do a file
transfer, go offline and online again and try to start another
file transfer.

Test-Information:

Tested that the mentioned use case does not crash anymore,
that file transfers still work, and that FileTransferTest still
passes.

Change-Id: Iddbcd8522af7df528bdc2b030fe616ad3f0c4701