Explicitly disconnect from remaining resources in destructors
Explicitly disconnect from remaining resources in the
ChainedConnected and ClientSessionStanzaChannel destructors, so
the event loop will not call the signal handler methods on a
freed object.

Test-Information:

Repeating the test case of creating a Swift::Client instance,
connecting it and then deleting it after a random time below one
second.

On Mac OS X 10.9.5 running this test case causes two ASAN
heap-use-after-free errors and with this patch the errors are
gone.

Change-Id: I3e48150c3633f4076ca9172aad9e85ba389df950