Disabling HiDPI support for Qt on Windows
This is an intermediate solution to the problem that the web views
of Qt do not scale to the correct size out of the box.

Disabling support for HiDPI mode will let Windows do pixel based
scaling. The resulting appearance is not as clear but the web
views remain readable.

Test-Information:

Tested with Qt 5.4.2 on Windows 8.

Change-Id: I9460902e728e6190c8382d4eb954135470654e5c