Add close button to about dialog for Linux desktops
The default Debian 7 desktop uses window decorations for dialog
windows that does not have close window buttons at the top. This
commit adds a dedicated close button to the about window.

This button is *not* added on Windows and OS X.

Test-Information:

Tested looks and function on OS X 10.9.5 with Qt 5.4.2 and
Debian 7 with Qt 5.3.2.

Change-Id: Ia363f66666eb88d43834ab0556f7b06efd3cedef