Only put qt.conf as resource binary for Windows builds
Commit 22ea5735ad1a105a4294b7475aba58646ef3ee8a added a qt.conf
file to the compiled binary resources of Swift which is picked up
by Qt and configures Qt to disable DPI awareness on Windows, as it
is not supported by the Qt WebKit widgets yet.

However, the Mac application bundle also ships with a qt.conf
file which set the correct folders for Qt plugins. This was
overridden by the compiled qt.conf resource file.

This commit only compiles the qt.conf resoruce file require for
Windows on the Windows platform.

Test-Information:

Tested on Windows 10 with 150% scaling that Swift is scaled up
based on the pixel data.
Tested on OS X 10.9.5 that the installation package works again.

Change-Id: I74ce722cff1090e7dafc3a67e04f2a45d375a43d