Remove Qt bearer plugins from Swift bundle in installation package
This is a workaround for QTBUG-34641. The Qt bearer API for
detecting network changes will poll for available wireless networks
every 30 seconds or so, resulting in needless CPU load.

This patch removes the plugins for this API from the installation
package, which will disable the constant polling for available
wireless networks.

Test-Information:

Verified that the plugins are removed from the application bundle
in the final installation DMG.

Change-Id: If46d03ca5623885879ecc10e06a29cec96e7caa4