Increase template OS X installation disk image size
Our packaging procedure temporarily needs more space on the disk
image as it first copies the application bundle and frameworks on
the image and afterwards removes debugging symbols and other files
not required by the end user.
The package.sh script reported this error of not having enough space
left on the disk image.

Test-Infromation:

Verified that the disk image is still build, the package.sh error
is gone and the packaged Swift.app works.

Change-Id: Iba853bd30eb63f74f07cdda08b93e76dd89368bb