Set keyboard focus to text input after selecting an emoticon
After selecting an emoticon on Windows the keyboard focus does not
return to the text input. OS X did not exhibit that behavior.

This fix explicitly sets the focus back to the text input after
selection of an emoticon, as expected by the user.

Test-Information:

Verified the behavior on OS X 10.9.5 and Windows 7, both with
Qt 5.4.2.

Change-Id: I8737a7e3f6d1a6b7a8e60261e9bd14c7be8d07a6