Save 'apply to all' property in highlight rule editor
The highlight editor did not save if the user set the 'apply to
all' property on a rule. It allows to highlight complete messages
or be notified only based on the incoming message type.
This commit enables saving the property.

Test-Information:

Tested chat and room rules with the 'apply to all' property set
and verified that the complete messages are highlighted.

Change-Id: Ib06bd3111ae810c8362ab8fd36a2d3b67f5a14cc