Do not require complete Qt installation when building Swiften
Previously, if a base installation of Qt was available but not
all Qt modules required by Swift are available Scons would fail
configuration. If a basic Qt installation is available SCons will
try to configure Swift even if you only want to build Swiften.

This commit changes the hard failure in Swift configuration to a
warning. This allows only Swiften to be build.

In addition, the warning now lists the Qt packages required by
Swift.

Test-Information:

Tested on Arch Linux.

Change-Id: I220ea61a7e4af849912e1fb1ae66cbecf5136d29