Keep scroll position in highlight rule editor when moving rules
When moving a rule up or down in the highlight rule editor on
Linux, the scroll area scrolled all the way down after each move.
This problem is not reproducible on OS X.

With this commit the dialog remembers the scroll area position
before moving a rule and sets the scroll position back to the
remembered value after the move of the highlight rule.

Test-Information:

Verified that the scroll position is not changed when moving
highlight rules up or down on Debian 8.2 and OS X 10.9.5.

Change-Id: Ide7c99e1311671c77cbf72da5cad4f64bfaab11f